RavenDB-18675 Clean up & const moved into Utils
