+++ src/Corax/Analyzers/Analyzer.cs
@@ -5,6 +5,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 using Corax.Pipeline;
+using Corax.Utils;
 
 namespace Corax.Analyzers
 {
