+++ src/Corax/IndexSearcher/IndexSearcher.TermMatch.cs
@@ -3,6 +3,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 using Corax.Queries;
+using Corax.Utils;
 using Sparrow.Server;
 using Sparrow.Server.Compression;
 using Voron;
