--- src/Corax/IndexingTableType.cs
@@ -1,8 +1,0 @@
-﻿namespace Corax
-{
-    public enum IndexingTableType : byte
-    {
-        None = 0,
-        IndexEntries = 1
-    }
-}
